Update to work on CentOS 7
The README is out of date for the version of OpenSearch being installed.
The README also does not provide sufficient information to successfully
execute the playbook.

Java 8 will not work and there are pre-requisite packages that are
needed.  The setup of these prerequisites are added to make the setup
more convenient.

The security setup was modified to wait on the OpenSearch process to
start seeing that the process takes a while to begin. This allows the
security setup to be successful where it might have failed because
OpenSearch was not started.